organize files in the codebase
